+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [3.5.0] - 2025-11-12
+
+### Added
+
+- Add support for uploading React Native Sourcemaps outside of Android or iOS builds [#241](https://github.com/bugsnag/bugsnag-cli/pull/241)
+
+### Changed
+
+- `bugsnag-cli-placeholder` now throws an `Error` instead of just logging to the console. This immediately stops execution and should return a stack trace if the CLI is not successfully installed [#239](https://github.com/bugsnag/bugsnag-cli/pull/239)
+
 ## [3.4.1] - 2025-09-30
 
 ### Fixed